keep .md files anywhere in the folder
@luizaandrade, thanks for catching this. I made some edits to your additions.

I changed the comment to the section where .gitignore etc. are not ignored. My previous comment was confusing as that section had nothing to do with the top folder from a ignore functionality perspective. After the edits in this commit, .md files are never ignored. Let me know if you see any issues with this.